@@ -1,6 +1,6 @@
-package java.com.baeldung.newfeatures;
+package com.baeldung.newfeatures;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -13,6 +13,6 @@ public class TeeingCollectorUnitTest {
     public void givenSetOfNumbers_thenCalculateAverage() {
         double mean = Stream.of(1, 2, 3, 4, 5)
           .collect(Collectors.teeing(Collectors.summingDouble(i -> i), Collectors.counting(), (sum, count) -> sum / count));
-        assertEquals(3.0, mean);
+        assertEquals(3.0, mean, 0);
     }
 }
