@@ -17,8 +17,8 @@ import java.nio.file.Paths;
 import java.security.NoSuchAlgorithmException;
 import java.time.Duration;
 
+import org.junit.Ignore;
 import org.junit.Test;
-import org.junit.jupiter.api.Disabled;
 
 public class HttpRequestUnitTest {
 
@@ -51,10 +51,10 @@ public class HttpRequestUnitTest {
 
     /*
     * This test will fail as soon as the given URL returns a HTTP 2 response.
-    * Therefore, let's leave it commented out.
+    * Therefore, let's leave it ignored.
     * */
-   @Test
-   @Disabled
+    @Test
+    @Ignore
     public void shouldFallbackToHttp1_1WhenWebsiteDoesNotUseHttp2() throws IOException, InterruptedException, URISyntaxException, NoSuchAlgorithmException {
         HttpRequest request = HttpRequest.newBuilder()
             .uri(new URI("https://postman-echo.com/get"))
