@@ -1,6 +1,8 @@
 package com.baeldung.uuid;
 
 import java.io.UnsupportedEncodingException;
+import java.nio.ByteOrder;
+import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.time.Duration;
@@ -132,21 +134,21 @@ public class UUIDGenerator {
     }
 
     private static byte[] bytesFromUUID(String uuidHexString) {
-        String normalizedUUIDHexString = uuidHexString.replace("-","");
+        String normalizedUUIDHexString = uuidHexString.replace("-", "");
 
         assert normalizedUUIDHexString.length() == 32;
 
         byte[] bytes = new byte[16];
         for (int i = 0; i < 16; i++) {
-            byte b = hexToByte(normalizedUUIDHexString.substring(i*2, i*2+2));
+            byte b = hexToByte(normalizedUUIDHexString.substring(i * 2, i * 2 + 2));
             bytes[i] = b;
         }
         return bytes;
     }
 
     public static byte hexToByte(String hexString) {
-        int firstDigit = Character.digit(hexString.charAt(0),16);
-        int secondDigit = Character.digit(hexString.charAt(1),16);
+        int firstDigit = Character.digit(hexString.charAt(0), 16);
+        int secondDigit = Character.digit(hexString.charAt(1), 16);
         return (byte) ((firstDigit << 4) + secondDigit);
     }
 
@@ -154,14 +156,47 @@ public class UUIDGenerator {
         int finalLength = byteArray1.length + byteArray2.length;
         byte[] result = new byte[finalLength];
 
-        for(int i = 0; i < byteArray1.length; i++) {
+        for (int i = 0; i < byteArray1.length; i++) {
             result[i] = byteArray1[i];
         }
 
-        for(int i = 0; i < byteArray2.length; i++) {
-            result[byteArray1.length+i] = byteArray2[i];
+        for (int i = 0; i < byteArray2.length; i++) {
+            result[byteArray1.length + i] = byteArray2[i];
         }
 
         return result;
+    }
+
+    public static UUID generateType5UUID(String name) {
+
+        try {
+
+            byte[] bytes = name.getBytes(StandardCharsets.UTF_8);
+            MessageDigest md = MessageDigest.getInstance("SHA-1");
+
+            byte[] hash = md.digest(bytes);
+
+            long msb = getLeastAndMostSignificantBitsVersion5(hash, 0);
+            long lsb = getLeastAndMostSignificantBitsVersion5(hash, 8);
+            // Set the version field
+            msb &= ~(0xfL << 12);
+            msb |= ((long) 5) << 12;
+            // Set the variant field to 2
+            lsb &= ~(0x3L << 62);
+            lsb |= 2L << 62;
+            return new UUID(msb, lsb);
+
+        } catch (NoSuchAlgorithmException e) {
+            throw new AssertionError(e);
+        }
+    }
+
+    private static long getLeastAndMostSignificantBitsVersion5(final byte[] src, final int offset) {
+        long ans = 0;
+        for (int i = offset + 7; i >= offset; i -= 1) {
+            ans <<= 8;
+            ans |= src[i] & 0xffL;
+        }
+        return ans;
     }
 }
