@@ -1,10 +1,10 @@
 package com.baeldung.metrics.micrometer;
 
+import io.micrometer.core.instrument.binder.jvm.JvmThreadMetrics;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
-import io.micrometer.core.instrument.binder.JvmThreadMetrics;
 
 @SpringBootApplication
 public class MicrometerApp {
@@ -14,7 +14,7 @@ public class MicrometerApp {
         return new JvmThreadMetrics();
     }
 
-    public static void main(String[] args) throws Exception {
+    public static void main(String[] args) {
         SpringApplication.run(MicrometerApp.class, args);
     }
 
