@@ -5,8 +5,8 @@ import com.mongodb.DBObject;
 import com.mongodb.client.MongoClients;
 import de.flapdoodle.embed.mongo.MongodExecutable;
 import de.flapdoodle.embed.mongo.MongodStarter;
-import de.flapdoodle.embed.mongo.config.IMongodConfig;
-import de.flapdoodle.embed.mongo.config.MongodConfigBuilder;
+import de.flapdoodle.embed.mongo.config.ImmutableMongodConfig;
+import de.flapdoodle.embed.mongo.config.MongodConfig;
 import de.flapdoodle.embed.mongo.config.Net;
 import de.flapdoodle.embed.mongo.distribution.Version;
 import de.flapdoodle.embed.process.runtime.Network;
@@ -36,7 +36,9 @@ class ManualEmbeddedMongoDbIntegrationTest {
         String ip = "localhost";
         int randomPort = SocketUtils.findAvailableTcpPort();
 
-        IMongodConfig mongodConfig = new MongodConfigBuilder().version(Version.Main.PRODUCTION)
+        ImmutableMongodConfig mongodConfig = MongodConfig
+            .builder()
+            .version(Version.Main.PRODUCTION)
             .net(new Net(ip, randomPort, Network.localhostIsIPv6()))
             .build();
 
