@@ -20,6 +20,7 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -30,7 +31,7 @@ import com.baeldung.web.exception.CustomException1;
 import com.baeldung.web.hateoas.event.PaginatedResultsRetrievedEvent;
 
 /**
- * 
+ *
  *  We'll start only the web layer.
  *
  */
@@ -54,20 +55,22 @@ public class FooControllerWebLayerIntegrationTest {
         doNothing().when(publisher)
             .publishEvent(any(PaginatedResultsRetrievedEvent.class));
 
-        this.mockMvc.perform(get("/foos").param("page", "0")
-            .param("size", "2"))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$",Matchers.hasSize(1)));
+        this.mockMvc.perform(get("/foos")
+          .param("page", "0")
+          .param("size", "2")
+          .accept(MediaType.APPLICATION_JSON))
+          .andExpect(status().isOk())
+          .andExpect(jsonPath("$", Matchers.hasSize(1)));
     }
-    
+
     @Test
     public void delete_forException_fromService() throws Exception {
         Mockito.when(service.findAll()).thenThrow(new CustomException1());
         this.mockMvc.perform(get("/foos")).andDo(h ->  {
             final Exception expectedException = h.getResolvedException();
             Assert.assertTrue(expectedException instanceof CustomException1);
-            
+
         });
     }
-    
+
 }
