@@ -25,7 +25,7 @@ public class FormAutoconfAuthenticationLiveTest {
         get(SVC_URL).then()
             .assertThat()
             .statusCode(HttpStatus.OK.value())
-            .content(containsString("<form"), containsString("action=\"login\""));
+            .body(containsString("<form"), containsString("action=\"login\""));
     }
 
     @Test
